Time: 2 ms (55.04%), Space: 8.4 MB (45.62%) - LeetHub

+++ 0007-reverse-integer/0007-reverse-integer.c
@@ -14,5 +14,5 @@ int reverse(int x) {
         rev_num = rev_num * 10 + digit;
     }
     
-    return (int)rev_num;
+    return rev_num;
 }
